% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmer.R
\name{mnmer}
\alias{mnmer}
\alias{cmnmer}
\alias{MNmer}
\title{mnmer}
\usage{
mnmer(seqs, m, n)
}
\arguments{
\item{seqs}{DNAStringSet object \code{seqs}}

\item{m}{Int value of m \code{m}}

\item{n}{Int value of n \code{n}}
}
\value{
Outputs a dataframe
}
\description{
Generates the feature matrix using conditional probability. As default, all sequences with N+IUPAC content higher 10\% than are removed.
}
\examples{
dir <-system.file("extdata", package="mnmer")
human <- readDNAStringSet(file.path(dir, "human_vir.fasta"))
human_02mer <- mnmer(human,2,0)

}
\keyword{mnmer}
\keyword{featureMatrix}
\keyword{k-mers}
