% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmer.R
\name{mnmer}
\alias{mnmer}
\alias{cmnmer}
\alias{MNmer}
\title{mnmer}
\usage{
mnmer(inputParameter1, inputParameter2, inputParameter3)
}
\arguments{
\item{inputParameter1}{FASTA file \code{inputParameter1}}

\item{inputParameter2}{Int value of k \code{inputParameter2}}

\item{inputParameter3}{Int value of m \code{inputParameter3}}
}
\value{
Outputs a dataframe
}
\description{
Generates the feature matrix using conditional probability.
}
\examples{
dir <-system.file("extdata", package="mnmer")
mosquito <- mnmer(file.path(dir, "mosquito_vir.fasta"),2,0)

}
\keyword{(m,n)-mer}
\keyword{featureMatrix}
\keyword{k-mers}
