% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmer.R
\name{mnmer}
\alias{mnmer}
\alias{cmnmer}
\alias{MNmer}
\title{mnmer}
\usage{
mnmer(FASTAfile, k, m, pni = 0.1)
}
\arguments{
\item{FASTAfile}{FASTA file \code{FASTAfile}}

\item{k}{Int value of k \code{k}}

\item{m}{Int value of m \code{m}}

\item{pni}{Cutoff percentage of N+IUPAC bases}
}
\value{
Outputs a dataframe
}
\description{
Generates the feature matrix using conditional probability.
}
\examples{
dir <-system.file("extdata", package="mnmer")
mosquito <- mnmer(file.path(dir, "mosquito_vir.fasta.gz"),2,0)

}
\keyword{(m,n)-mer}
\keyword{featureMatrix}
\keyword{k-mers}
