% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmer.R
\name{mnmer}
\alias{mnmer}
\title{mnmer}
\usage{
mnmer(file, k, m)
}
\arguments{
\item{inputParameter1}{FASTA file \code{inputParameter1}}

\item{inputParameter2}{Int value of k \code{inputParameter2}}

\item{inputParameter3}{Int value of m \code{inputParameter3}}
}
\value{
Outputs a dataframe
}
\description{
Generates the feature matrix using conditional probability.
}
\examples{
features <- mnmer(sequences.fasta,2,0)
features <- mnmer(sequences.fasta,2,1)

}
\keyword{(m,n)-mer}
\keyword{featureMatrix}
\keyword{k-mers}
