% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNumFASTA.R
\name{readNumFASTA}
\alias{readNumFASTA}
\alias{readrandFASTA}
\alias{readNum}
\alias{readFASTA}
\title{readNumFASTA}
\usage{
readNumFASTA(FASTAfile, size = 0, rand = FALSE, pni = 0.2)
}
\arguments{
\item{FASTAfile}{Path to a multifasta file \code{FASTAfile}}

\item{size}{Number of sequences to be loaded \code{size}}

\item{rand}{Sequences choose mode random or not. TRUE or FALSE \code{rand}}

\item{pni}{Cutoff percentage for maximum of non-ACTG bases in the sequences \code{pni}}
}
\value{
DNAStringSet object
}
\description{
Load data in to the R system.
}
\examples{
dir <-system.file("extdata", package="mnmer")
human <-readNumFASTA((file.path(dir, "human_vir.fasta")), 10, TRUE, 0.50)

}
\keyword{non-ACTG}
\keyword{random}
\keyword{sequences}
